requirements.txt
python-telegram-bot
openpyxl
beautifulsoup4
requests